Refactor move migrations to pks from pgmodel
